fix bug

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,7 +3,7 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
         int fahrenheit;
-        fahrenheit = (temperatureCelsius * 1.8) + 32;
+        fahrenheit = (temperatureCelsius * (int)1.8) + 32;
         System.out.println(fahrenheit);
     }
 }
